requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
newsapi-python==0.2.6
requests==2.31.0
textblob==0.17.1
streamlit==1.25.0
python-dotenv==1.0.0
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2 